feat/refactor: login page, app refactor
